docs(Flow): samples refactored
