+++ packages/lab/src/components/Flow/stories/Base/Dashboard.tsx
@@ -1,4 +1,4 @@
-import { HvFlowNode } from "../Node/Node";
+import { HvFlowNode } from "@hitachivantara/uikit-react-lab";
 
 export const Dashboard = (props) => {
   return (
