+++ packages/lab/src/components/Flow/stories/Base/MLModelDetection.tsx
@@ -1,4 +1,4 @@
-import { HvFlowNode } from "../Node/Node";
+import { HvFlowNode } from "@hitachivantara/uikit-react-lab";
 
 export const MLModelDetection = (props) => {
   return <HvFlowNode description="Anomaly detection description" {...props} />;
