+++ packages/lab/src/components/Flow/stories/Visualizations/JsonInput.tsx
@@ -1,4 +1,4 @@
-import { HvFlowNode } from "../../Node/Node";
+import { HvFlowNode } from "@hitachivantara/uikit-react-lab";
 
 export const JsonInput = (props) => {
   return <HvFlowNode description="Population Datakky7" {...props} />;
