+++ packages/lab/src/components/Flow/stories/Visualizations/LineChart.tsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/css";
+import { HvFlowNode } from "@hitachivantara/uikit-react-lab";
 import { HvLineChart } from "@hitachivantara/uikit-react-viz";
 import { useStore } from "reactflow";
-import { HvFlowNode } from "../../Node/Node";
 
 export const LineChart = (props) => {
   const { id } = props;
