+++ packages/lab/src/components/types/global.d.ts
@@ -1,0 +1,3 @@
+declare module "*?raw" {
+  export default "" as string;
+}
